Reimplement hook_NtCreateFile to allow it to create reroutes
