+++ src/shared/ntdll_declarations.h
@@ -141,6 +141,7 @@ typedef struct _FILE_REPARSE_POINT_INFORMATION {
   ULONG Tag;
 } FILE_REPARSE_POINT_INFORMATION, *PFILE_REPARSE_POINT_INFORMATION;
 
+// copied from ntstatus.h
 #define STATUS_SUCCESS ((NTSTATUS)0x00000000L)
 #define STATUS_BUFFER_OVERFLOW ((NTSTATUS)0x80000005L)
 #define STATUS_NO_MORE_FILES ((NTSTATUS)0x80000006L)
